requirements.txt
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
joblib
scikit-learn
requests
python-dotenv
beautifulsoup4
pandas